suppress compiler warnings
git-svn-id: https://svn.r-project.org/R/trunk@79165 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/connections.c
@@ -437,7 +437,7 @@ int dummy_vfprintf(Rconnection con, const char *format, va_list ap)
 	res = vasprintf(&b, format, ap);
 	if (res < 0) {
 	    b = buf;
-	    res = strlen(buf) + 1;
+	    res = (int)strlen(buf) + 1;
 	    buf[res-1] = '\n';
 	    warning(_("printing of extremely long output is truncated"));
 	} else usedVasprintf = TRUE;
